refactor: rename page/ to pages/
